requirements.txt
pyside6
PyQt6-tools
python-dotenv
mysql-connector-python